Move abstract constraint classes
