+++ src/model/constraints/constraint.py
@@ -1,4 +1,4 @@
-from abc import ABC
+from abc import ABC, abstractmethod
 
 # TODO some work here and in specification.py to clarify the conceptual abstrcation levels (what is meta, what is instance)
 
@@ -17,3 +17,16 @@ class Constraint(ABC):
 
     def __repr__(self):
         return f"{self.predicate_name}{self.actors}"
+
+class Behavior_Con(Constraint):
+    @abstractmethod
+    def get_possible_paths(self): # TODO: Confirm naming and implement
+        return []
+
+class Static_Con(Constraint):
+    @abstractmethod
+    def get_heuristic_value(self):
+        pass
+
+    def is_satisfied(self):
+        return self.get_heuristic_value() == 0
